殷海薩調整 (修復登入時的顯示) 避免多餘的宣告 避免多餘的轉型
git-svn-id: http://www.castiello-tsai.org:8523/svn/L1JTW@1704 96aee97a-5c11-ed42-8518-d7f33e96f5f9

+++ src/net/l1j/server/serverpackets/S_FollowerPack.java
@@ -25,14 +25,14 @@ import net.l1j.server.model.instance.L1FollowerInstance;
 public class S_FollowerPack extends ServerBasePacket {
 	private static final String S_FOLLOWER_PACK = "[S] S_FollowerPack";
 
-	private static final short STATUS_POISON = 1;
-	private static final short STATUS_INVISIBLE = 2;
-	private static final short STATUS_PC = 4;
-	private static final short STATUS_FREEZE = 8;
-	private static final short STATUS_BRAVE = 16;
-	private static final short STATUS_ELFBRAVE = 32;
-	private static final short STATUS_FASTMOVABLE = 64;
-	private static final short STATUS_GHOST = 128;
+	private static final int STATUS_POISON = 1;
+	private static final int STATUS_INVISIBLE = 2;
+	private static final int STATUS_PC = 4;
+	private static final int STATUS_FREEZE = 8;
+	private static final int STATUS_BRAVE = 16;
+	private static final int STATUS_ELFBRAVE = 32;
+	private static final int STATUS_FASTMOVABLE = 64;
+	private static final int STATUS_GHOST = 128;
 
 	private byte[] _byte = null;
 
@@ -50,7 +50,7 @@ public class S_FollowerPack extends ServerBasePacket {
 		writeH(0);
 		writeS(follower.getNameId());
 		writeS(follower.getTitle());
-		short status = 0;
+		int status = 0;
 		if (follower.getPoison() != null) { // 毒狀態
 			if (follower.getPoison().getEffectId() == 1) {
 				status |= STATUS_POISON;
