+++ src/net/l1j/server/serverpackets/S_OtherCharPacks.java
@@ -24,16 +24,16 @@ import net.l1j.server.model.instance.L1PcInstance;
 public class S_OtherCharPacks extends ServerBasePacket {
 	private static final String S_OTHER_CHAR_PACKS = "[S] S_OtherCharPacks";
 
-	private static final short STATUS_POISON = 1;
-	private static final short STATUS_INVISIBLE = 2;
-	private static final short STATUS_PC = 4;
-	private static final short STATUS_FREEZE = 8;
-	private static final short STATUS_BRAVE = 16;
-	private static final short STATUS_ELFBRAVE = 32;
-	private static final short STATUS_FASTMOVABLE = 64;
-	private static final short STATUS_GHOST = 128;
-	private static final short STATUS_EXPBLESS = 256;
-	private static final short STATUS_TRIPLES_SPEED = 512;
+	private static final int STATUS_POISON = 1;
+	private static final int STATUS_INVISIBLE = 2;
+	private static final int STATUS_PC = 4;
+	private static final int STATUS_FREEZE = 8;
+	private static final int STATUS_BRAVE = 16;
+	private static final int STATUS_ELFBRAVE = 32;
+	private static final int STATUS_FASTMOVABLE = 64;
+	private static final int STATUS_GHOST = 128;
+	private static final int STATUS_EXPBLESS = 256;
+	private static final int STATUS_TRIPLES_SPEED = 512;
 
 	private byte[] _byte = null;
 
@@ -47,7 +47,7 @@ public class S_OtherCharPacks extends ServerBasePacket {
 	}
 
 	private void buildPacket(L1PcInstance pc, boolean isFindInvis) {
-		short status = STATUS_PC;
+		int status = STATUS_PC;
 
 		if (pc.getPoison() != null) { // 毒狀態
 			if (pc.getPoison().getEffectId() == 1) {
